Code/WirelessDebugger/requirements.txt
pyqtgraph==0.12.2
numpy==1.20.3
PyQt5==5.15.6
pyserial==3.5
